// src/types/nfe.ts
export type ParsedNFeItem = {
    codigo: string;
    descricao: string;
    ncm: string;
    cest?: string;
    unidade: string;
    quantidade: number;
    valor_unitario: number;
    valor_total: number;
    valor_frete?: number;
    valor_seguro?: number;
    valor_desconto?: number;
    valor_outras_despesas?: number;
    origem?: string;
    cfop?: string;
    csosn?: string;
    cbenef?: string;
    ipi_cst?: string;
    ipi_cenq?: string;
    ipi_base?: number;
    ipi_aliquota?: number;
    ipi_valor?: number;
    pis_cst?: string;
    pis_base?: number;
    pis_aliquota?: number;
    pis_valor?: number;
    cofins_cst?: string;
    cofins_base?: number;
    cofins_aliquota?: number;
    cofins_valor?: number;
};